+++ src/client/http/requestmanager/RequestManager.h
@@ -3,6 +3,7 @@
 #include "common/String.h"
 #include "client/http/PostData.h"
 #include <atomic>
+#include <cstdint>
 #include <thread>
 #include <vector>
 #include <memory>
@@ -38,8 +39,8 @@ namespace http
 		State state = ready;
 		std::mutex stateMx;
 		std::condition_variable stateCv;
-		std::atomic<int> bytesTotal = -1;
-		std::atomic<int> bytesDone = 0;
+		std::atomic<int64_t> bytesTotal = -1;
+		std::atomic<int64_t> bytesDone = 0;
 		int statusCode = 0;
 		ByteString responseData;
 		std::vector<ByteString> responseHeaders;
